chore(ui-ux): refactor poolpair icon component
