+++ mobile-app/app/screens/AppNavigator/screens/Dex/components/PoolPairCards/PoolPairIconV2.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { StyleProp, View, ViewProps } from 'react-native'
+import { tailwind } from '@tailwind'
+import { getNativeIcon } from '@components/icons/assets'
+
+export function PoolPairIconV2 (props: {
+  symbolA: string
+  symbolB: string
+  customSize?: number
+  iconBStyle?: StyleProp<ViewProps>
+}): JSX.Element {
+  const IconA = getNativeIcon(props.symbolA)
+  const IconB = getNativeIcon(props.symbolB)
+  return (
+    <View style={tailwind('flex-row')}>
+      <IconA height={props.customSize ?? 40} width={props.customSize ?? 40} style={tailwind('relative z-10')} />
+      <IconB height={props.customSize ?? 40} width={props.customSize ?? 40} style={[tailwind('-ml-4 mr-2'), props.iconBStyle]} />
+    </View>
+  )
+}
